requirements.txt
h5py
torch
torchvision
numpy
pandas
matplotlib
tqdm
pyyaml
scikit-learn
tensorboard
librosa
fastapi
uvicorn
python-multipart
pytest